Refactor chat input into its own component
